feat(pitch-challenge): Add Mimic Game navigation and update pitch detection
Introduces navigation to the Mimic Game screen from HomeScreen and AppNavGraph, including a new HomeCard for the Humming Challenge. Also fixes pitch detection note calculation in PitchDetector. SoundGenerator.kt was moved to the utils package for better organization.

+++ app/src/main/java/np/ict/mad/mad25_p03_team03/utils/PitchDetector.kt
@@ -51,7 +51,7 @@ class PitchDetector {
         if (hz < 20) return "" // 噪音过滤
         val notes = arrayOf("C", "C#", "D", "D#", "E", "F", "F#", "G", "G#", "A", "A#", "B")
         val log2 = Math.log(hz / 440.0) / Math.log(2.0)
-        val noteNumber = Math.round(12 * log2) + 69 // MIDI note number
+        val noteNumber = (Math.round(12 * log2) + 69).toInt() // MIDI note number
         val octave = (noteNumber / 12) - 1
         val noteIndex = noteNumber % 12
         return if (noteIndex >= 0 && noteIndex < notes.size) {
